Inline transverse mercator implementation.
Since there are no immediate plans to expose generic transverse projections,
it’s slightly faster and more concise to inline this implementation.
